Add common.csproj.props to manage project attributes better
