+++ src/CromulentBisgetti.ContainerPacking/CromulentBisgetti.ContainerPacking.csproj
@@ -1,7 +1,6 @@
 <Project Sdk="Microsoft.NET.Sdk">
-
+  <Import Project="$(RepositoryRootDir)common.csproj.props" />
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
   </PropertyGroup>
-
 </Project>
