+++ src/CromulentBisgetti.ContainerPacking/Entities/Item.cs
@@ -23,7 +23,7 @@ namespace CromulentBisgetti.ContainerPacking.Entities
 		/// <param name="dim1">The length of one of the three item dimensions.</param>
 		/// <param name="dim2">The length of another of the three item dimensions.</param>
 		/// <param name="dim3">The length of the other of the three item dimensions.</param>
-		/// <param name="itemQuantity">The item quantity.</param>
+		/// <param name="quantity">The item quantity.</param>
 		public Item(int id, decimal dim1, decimal dim2, decimal dim3, int quantity)
 		{
 			this.ID = id;
