--- src/CromulentBisgetti.ContainerPackingTests/CromulentBisgetti.ContainerPackingTests.csproj
@@ -1,27 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
-
-    <IsPackable>false</IsPackable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <None Remove="DataFiles\ORLibrary.txt" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <EmbeddedResource Include="DataFiles\ORLibrary.txt" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.8.0" />
-    <PackageReference Include="MSTest.TestAdapter" Version="1.3.2" />
-    <PackageReference Include="MSTest.TestFramework" Version="1.3.2" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\CromulentBisgetti.ContainerPacking\CromulentBisgetti.ContainerPacking.csproj" />
-  </ItemGroup>
-
-</Project>
