+++ src/CromulentBisgetti.DemoApp/CromulentBisgetti.DemoApp.csproj
@@ -1,5 +1,5 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
-
+  <Import Project="$(RepositoryRootDir)common.csproj.props" />
   <PropertyGroup>
     <TargetFramework>netcoreapp2.1</TargetFramework>
   </PropertyGroup>
